# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/dynamic/experiment-layer-bindings/create-bind": {
            "post": {
                "description": "创建一个Experiment与Layer的绑定关系",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "绑定关系"
                ],
                "summary": "创建Experiment与Layer绑定关系",
                "parameters": [
                    {
                        "description": "绑定信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb_dynamic.CreateExperimentLayerBindingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb_dynamic.CreateExperimentLayerBindingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/dynamic/experiment-layer-bindings/delete-bind": {
            "delete": {
                "description": "删除特定的Experiment与Layer绑定关系",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "绑定关系"
                ],
                "summary": "删除Experiment与Layer绑定关系",
                "parameters": [
                    {
                        "description": "删除请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb_dynamic.DeleteExperimentLayerBindingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb_dynamic.DeleteExperimentLayerBindingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/dynamic/experiment-layer-bindings/get-bind": {
            "get": {
                "description": "获取特定的Experiment与Layer绑定关系",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "绑定关系"
                ],
                "summary": "获取Experiment与Layer绑定关系",
                "parameters": [
                    {
                        "type": "string",
                        "name": "experiment_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "layer_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb_dynamic.GetExperimentLayerBindingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/dynamic/experiment-user-bindings/create-bind": {
            "post": {
                "description": "创建一个Experiment与User的绑定关系",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "绑定关系"
                ],
                "summary": "创建Experiment与User绑定关系",
                "parameters": [
                    {
                        "description": "绑定信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb_dynamic.CreateExperimentUserBindingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb_dynamic.CreateExperimentUserBindingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/dynamic/experiment-user-bindings/delete-bind": {
            "delete": {
                "description": "删除特定的Experiment与User绑定关系",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "绑定关系"
                ],
                "summary": "删除Experiment与User绑定关系",
                "parameters": [
                    {
                        "description": "删除请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb_dynamic.DeleteExperimentUserBindingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb_dynamic.DeleteExperimentUserBindingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/dynamic/experiment-user-bindings/get-bind": {
            "get": {
                "description": "获取特定的Experiment与User绑定关系",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "绑定关系"
                ],
                "summary": "获取Experiment与User绑定关系",
                "parameters": [
                    {
                        "type": "string",
                        "name": "experiment_group_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb_dynamic.GetExperimentUserBindingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/dynamic/layer-experiments/get-all-bylayer": {
            "get": {
                "description": "获取指定Layer关联的所有实验信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "实验"
                ],
                "summary": "获取Layer关联的所有实验",
                "parameters": [
                    {
                        "type": "string",
                        "name": "layer_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb_dynamic.GetLayerExperimentsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/dynamic/layer-user-bindings/create-bind": {
            "post": {
                "description": "创建一个Layer与User的绑定关系",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "绑定关系"
                ],
                "summary": "创建Layer与User绑定关系",
                "parameters": [
                    {
                        "description": "绑定信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb_dynamic.CreateLayerUserBindingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb_dynamic.CreateLayerUserBindingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/dynamic/layer-user-bindings/delete-bind": {
            "delete": {
                "description": "删除特定的Layer与User绑定关系",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "绑定关系"
                ],
                "summary": "删除Layer与User绑定关系",
                "parameters": [
                    {
                        "description": "删除请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb_dynamic.DeleteLayerUserBindingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb_dynamic.DeleteLayerUserBindingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/dynamic/layer-user-bindings/get-bind": {
            "get": {
                "description": "获取特定的Layer与User绑定关系",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "绑定关系"
                ],
                "summary": "获取Layer与User绑定关系",
                "parameters": [
                    {
                        "type": "string",
                        "name": "layer_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb_dynamic.GetLayerUserBindingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "pb_dynamic.CreateExperimentLayerBindingRequest": {
            "type": "object",
            "properties": {
                "binding": {
                    "$ref": "#/definitions/pb_dynamic.ExperimentLayerBinding"
                }
            }
        },
        "pb_dynamic.CreateExperimentLayerBindingResponse": {
            "type": "object",
            "properties": {
                "binding": {
                    "$ref": "#/definitions/pb_dynamic.ExperimentLayerBinding"
                }
            }
        },
        "pb_dynamic.CreateExperimentUserBindingRequest": {
            "type": "object",
            "properties": {
                "binding": {
                    "$ref": "#/definitions/pb_dynamic.ExperimentUserBinding"
                }
            }
        },
        "pb_dynamic.CreateExperimentUserBindingResponse": {
            "type": "object",
            "properties": {
                "binding": {
                    "$ref": "#/definitions/pb_dynamic.ExperimentUserBinding"
                }
            }
        },
        "pb_dynamic.CreateLayerUserBindingRequest": {
            "type": "object",
            "properties": {
                "binding": {
                    "$ref": "#/definitions/pb_dynamic.LayerUserBinding"
                }
            }
        },
        "pb_dynamic.CreateLayerUserBindingResponse": {
            "type": "object",
            "properties": {
                "binding": {
                    "$ref": "#/definitions/pb_dynamic.LayerUserBinding"
                }
            }
        },
        "pb_dynamic.DeleteExperimentLayerBindingRequest": {
            "type": "object",
            "properties": {
                "experiment_id": {
                    "type": "string"
                },
                "layer_id": {
                    "type": "string"
                }
            }
        },
        "pb_dynamic.DeleteExperimentLayerBindingResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                }
            }
        },
        "pb_dynamic.DeleteExperimentUserBindingRequest": {
            "type": "object",
            "properties": {
                "experiment_group_id": {
                    "type": "string"
                }
            }
        },
        "pb_dynamic.DeleteExperimentUserBindingResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                }
            }
        },
        "pb_dynamic.DeleteLayerUserBindingRequest": {
            "type": "object",
            "properties": {
                "layer_id": {
                    "type": "string"
                }
            }
        },
        "pb_dynamic.DeleteLayerUserBindingResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                }
            }
        },
        "pb_dynamic.ExperimentLayerBinding": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "创建时间 (Unix时间戳)",
                    "type": "integer"
                },
                "experiment_id": {
                    "description": "实验ID",
                    "type": "string"
                },
                "layer_id": {
                    "description": "层ID",
                    "type": "string"
                },
                "ratio": {
                    "description": "占有比例",
                    "type": "number"
                },
                "updated_at": {
                    "description": "更新时间 (Unix时间戳)",
                    "type": "integer"
                }
            }
        },
        "pb_dynamic.ExperimentUserBinding": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "创建时间 (Unix时间戳)",
                    "type": "integer"
                },
                "experiment_group_id": {
                    "description": "实验组ID",
                    "type": "string"
                },
                "updated_at": {
                    "description": "更新时间 (Unix时间戳)",
                    "type": "integer"
                },
                "user_package_id": {
                    "description": "用户包ID",
                    "type": "string"
                }
            }
        },
        "pb_dynamic.GetExperimentLayerBindingResponse": {
            "type": "object",
            "properties": {
                "binding": {
                    "$ref": "#/definitions/pb_dynamic.ExperimentLayerBinding"
                }
            }
        },
        "pb_dynamic.GetExperimentUserBindingResponse": {
            "type": "object",
            "properties": {
                "binding": {
                    "$ref": "#/definitions/pb_dynamic.ExperimentUserBinding"
                }
            }
        },
        "pb_dynamic.GetLayerExperimentsResponse": {
            "type": "object",
            "properties": {
                "bindings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pb_dynamic.ExperimentLayerBinding"
                    }
                },
                "remaining_ratio": {
                    "description": "剩余可用流量比例",
                    "type": "number"
                }
            }
        },
        "pb_dynamic.GetLayerUserBindingResponse": {
            "type": "object",
            "properties": {
                "binding": {
                    "$ref": "#/definitions/pb_dynamic.LayerUserBinding"
                }
            }
        },
        "pb_dynamic.LayerUserBinding": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "创建时间 (Unix时间戳)",
                    "type": "integer"
                },
                "layer_id": {
                    "description": "层ID",
                    "type": "string"
                },
                "updated_at": {
                    "description": "更新时间 (Unix时间戳)",
                    "type": "integer"
                },
                "user_package_id": {
                    "description": "用户包ID",
                    "type": "string"
                }
            }
        },
...
    }
}